common: Refactor cloud common to use internal common functions
Related: RHEL-60148
